refactor: split docs pure member rendering
